Rename hazmat -> danger
